Implement the ability to add item offers w/out photos for now

+++ shop/models.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User
 class Item(models.Model):
     name = models.CharField(max_length=255, default='')
     description = models.TextField(default='')
-    price = models.DecimalField(max_digits=4, decimal_places=2)
+    price = models.FloatField()
     category = models.TextField(default='')
 
     def __str__(self):
